// src/app/page/master/eselon-form/eselon-form.interface.ts
import {EventEmitter} from '@angular/core';

export interface Eselon {
    kodeEselon?:string;
	namaEselon?:string;
	reportDisplay?:string;
	kodeExternal?:string;
	namaExternal?:string;
	statusEnabled?: boolean;
	id?: any;
	kdProfile?: string;
	kode?: any;
	version?: any;
	label?: any;
	Version?:any;
	id_kode?: any;
}